Extract ui-fulldeps expression parser into module
